% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockTable.R
\name{stockTable}
\alias{stockTable}
\title{查看即时库存更新报表}
\usage{
stockTable(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
查看即时库存更新报表
}
\examples{
stockTable()
}
